Gestion des Contacts / Last commit !

+++ hoptest/app/Http/Controllers/contactController.php
@@ -6,7 +6,6 @@ use App\Contact;
 use App\Http\Validation\contactsValidator;
 use App\Http\Validation\ContactValidator;
 use Illuminate\Http\Request;
-use App\Http\Validation\ContactValidation;
 use Illuminate\Support\Facades\Validator;
 
 class contactController extends Controller
@@ -26,45 +25,37 @@ class contactController extends Controller
     public function delete_Contact($id)
     {
         $contact = Contact::find($id);
-        if($contact)
-            $contact->delete();
-
+        if ($contact) {$contact->delete();}
     }
 
-    public function find_by_id($id){
-        $contact = Contact::find($id);
-
-    }
-
-    public function edit_Contact(request $request, contactsValidator $validation,$id)
+    public function find_by_id($id)
     {
-        $validator = Validator::make($request->all(),$validation->rules(),$validation->messages());
+        $contact = Contact::find($id);
+    }
+
+    public function edit_Contact(request $request, contactsValidator $validation, $id)
+    {
+        $validator = Validator::make($request->all(), $validation->rules(), $validation->messages());
         $contact = Contact::find($id);
 
-        if($validator->fails())
-            return view('contacts/modifierContact', compact('contact'));
+        if ($validator->fails()) {return view('contacts/modifierContact', compact('contact'));}
 
+        $contact->nom               = $request->nom;
+        $contact->prenom            = $request->prenom;
+        $contact->fonction          = $request->fonction;
+        $contact->service           = $request->service;
+        $contact->date_de_naissance = date("Y-m-d", strtotime($request->datedenaissance));
+        $contact->email             = $request->email;
+        $contact->tel               = $request->tel;
+        $contact->nom_societe       = $request->nomsociete;
+        $contact->adresse           = $request->adresse;
+        $contact->ville             = $request->ville;
+        $contact->code_postal       = $request->cp;
+        $contact->site_web          = $request->siteweb;
 
-         $contact->nom = $request->nom ;
-        $contact->prenom = $request->prenom ;
-        $contact->fonction = $request->fonction ;
-        $contact->service = $request->service ;
-        $contact->date_de_naissance=  date("Y-m-d",strtotime($request->datedenaissance));
-        //$contact->date_de_naissance = str_replace("/",'-',$request->datedenaissance) ;
-        $contact->email = $request->email ;
-        $contact->tel = $request->tel ;
-        $contact->nom_societe = $request->nomsociete ;
-        $contact->adresse = $request->adresse ;
-        $contact->ville = $request->ville ;
-        $contact->code_postal = $request->cp ;
-        $contact->site_web = $request->siteweb ;
-
-        $contact -> save();
+        $contact->save();
         $contacts = Contact::all();
         return view('contacts/listeDesContacts', compact('contacts'));
-
-
-
     }
 
 }
